pyproject.toml
[tool.poetry]
name = "notifications-app"
version = "0.1.0"
description = ""
authors = ["chugunova24 <t.chugunova24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.2.3"
bcrypt = "^4.0.1"
flask-restful = "^0.3.10"
mongoengine = "^0.27.0"
flask-mongoengine = "^1.0.0"
flask-wtf = "1.1.1"
werkzeug = "2.2.2"
python-dotenv = "^1.0.0"
blinker = "^1.6.3"
marshmallow = "^3.20.1"
flask-mail = "^0.9.1"
celery = "5.2.6"
redis = "^5.0.1"
flask-celeryext = "^0.5.0"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
